# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " uala twitter-like API",
        "description": "Documentación de la API generada con Swagger/OpenAPI",
        "contact": {
            "email": "estebandiaczun@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/timeline": {
            "get": {
                "tags": [
                    "Timeline"
                ],
                "summary": "Obtener el timeline del usuario",
                "description": "Obtiene los tweets más recientes de los usuarios seguidos, ordenados cronológicamente. El timeline se actualiza en tiempo real cuando los usuarios que sigues publican nuevos tweets.",
                "operationId": "getTimeline",
                "parameters": [
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "description": "ID del usuario autenticado, enviado como encabezado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación. El valor predeterminado es 1.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de tweets por página. El valor predeterminado es 20.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timeline obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "tweet_id": {
                                                        "type": "string",
                                                        "example": "507f1f77bcf86cd799439011"
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "123e4567-e89b-12d3-a456-426614174000"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "¡Este es mi nuevo tweet!"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-20T06:12:27+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-User-ID": []
                    }
                ]
            }
        },
...
    },
...
}